backend/apps/billing/tests.py
import pytest
from django.contrib.auth import get_user_model
from rest_framework.test import APIClient
from rest_framework import status

User = get_user_model()


@pytest.mark.django_db
class TestBillingViews:
    """Test billing-related endpoints"""

    def test_billing_endpoint_placeholder(self, authenticated_client):
        """Placeholder test for billing endpoints"""
        # Add tests when billing functionality is implemented
        assert True


@pytest.mark.django_db
class TestBillingModels:
    """Test billing models"""

    def test_billing_model_placeholder(self):
        """Placeholder test for billing models"""
        # Add tests when billing models are implemented
        assert True


