backend/requirements/development.txt
-r base.txt
django-debug-toolbar>=4.0.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0

